+++ tests/message-extensions/src/index.ts
@@ -1,3 +1,5 @@
+import path from 'path';
+
 import { cardAttachment } from '@microsoft/teams.api';
 import { App } from '@microsoft/teams.apps';
 import { IAdaptiveCard } from '@microsoft/teams.cards';
@@ -11,6 +13,7 @@ import {
   createLinkUnfurlCard,
   createMessageDetailsCard,
 } from './card';
+
 const app = new App({
   logger: new ConsoleLogger('@tests/message-extensions', { level: 'debug' }),
   plugins: [new DevtoolsPlugin()],
@@ -128,6 +131,68 @@ app.on('message.ext.query', async ({ activity }) => {
   return { status: 400 };
 });
 // :snippet-end: message-ext-query
+
+// :snippet-start: message-ext-select-item
+app.on('message.ext.select-item', async ({ activity, send }) => {
+  const { option } = activity.value;
+
+  await send(`Selected item: ${option}`);
+
+  return {
+    status: 200,
+  };
+});
+// :snippet-end: message-ext-select-item
+
+// :snippet-start: message-ext-query-settings-url
+app.on('message.ext.query-settings-url', async ({ activity }) => {
+  // Get user settings from storage if available
+  const userSettings = await app.storage.get(activity.from.id) || { selectedOption: '' };
+  const escapedSelectedOption = encodeURIComponent(userSettings.selectedOption);
+
+  return {
+    composeExtension: {
+      type: 'config',
+      suggestedActions: {
+        actions: [
+          {
+            type: 'openUrl',
+            title: 'Settings',
+            // ensure the bot endpoint is set in the environment variables
+            // process.env.BOT_ENDPOINT is not populated by default in the Teams Toolkit setup. 
+            value: `${process.env.BOT_ENDPOINT}/tabs/settings?selectedOption=${escapedSelectedOption}`
+          }
+        ]
+      }
+    }
+  };
+});
+// :snippet-end: message-ext-query-settings-url
+
+// :snippet-start: message-ext-setting
+app.on('message.ext.setting', async ({ activity, send }) => {
+  const { state } = activity.value;
+  if (state == 'CancelledByUser') {
+    return {
+      status: 400
+    };
+  }
+  const selectedOption = state;
+  
+  // Save the selected option to storage
+  await app.storage.set(activity.from.id, { selectedOption });
+  
+  await send(`Selected option: ${selectedOption}`);
+
+  return {
+    status: 200
+  };
+});
+// :snippet-end: message-ext-setting
+
+// :snippet-start: message-ext-serve-html
+app.tab('settings', path.resolve(__dirname));
+// :snippet-end: message-ext-serve-html
 
 (async () => {
   await app.start();