+++ packages/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @microsoft/teams.common
 
+## 2.0.0-preview.6
+
+### Patch Changes
+
+- Disable automatic function calling, and show structured output
+
+## 2.0.0-preview.5
+
+### Patch Changes
+
+- streaming bug fix
+
 ## 2.0.0-preview.4
 
 ### Patch Changes