+++ tests/message-extensions/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @tests/message-extensions
 
+## 0.0.1-preview.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @microsoft/teams.api@2.0.0-preview.6
+  - @microsoft/teams.apps@2.0.0-preview.6
+  - @microsoft/teams.cards@2.0.0-preview.6
+  - @microsoft/teams.common@2.0.0-preview.6
+  - @microsoft/teams.dev@2.0.0-preview.6
+  - @microsoft/teams.graph@2.0.0-preview.6
+
+## 0.0.1-preview.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @microsoft/teams.apps@2.0.0-preview.5
+  - @microsoft/teams.api@2.0.0-preview.5
+  - @microsoft/teams.cards@2.0.0-preview.5
+  - @microsoft/teams.common@2.0.0-preview.5
+  - @microsoft/teams.dev@2.0.0-preview.5
+  - @microsoft/teams.graph@2.0.0-preview.5
+
 ## 0.0.1-preview.4
 
 ### Patch Changes